rename button component
