+++ app/src/components/NoData.vue
@@ -2,12 +2,12 @@
   <div class="flex flex-col items-center flex-grow">
     <i-fa-solid-ban class="text-3xl mt-auto" />
     <div class="flex flex-col items-center my-2 text-lg"><slot /></div>
-    <CustomButton class="mt-auto mb-4" :to="{ name: 'home' }" replace>
+    <Button class="mt-auto mb-4" :to="{ name: 'home' }" replace>
       <i-fa-close class="mr-2" /><span>Schließen</span>
-    </CustomButton>
+    </Button>
   </div>
 </template>
 
 <script setup lang="ts">
-import CustomButton from '~/components/atomic/CustomButton.vue';
+import Button from '~/components/atomic/Button.vue';
 </script>
