+++ app/src/components/popups/AboutPopup.vue
@@ -10,12 +10,12 @@
         deinem Bus an.
       </p>
       <p class="mb-2">Unterstütze uns indem du die App mit deinen Freunden teilst und gib uns Feedback!</p>
-      <CustomButton href="https://www.instagram.com/kiel.live/" class="w-full">
+      <Button href="https://www.instagram.com/kiel.live/" class="w-full">
         <i-mdi-instagram class="mr-2" /><span>Follow @kiel.live</span>
-      </CustomButton>
-      <CustomButton href="https://github.com/kiel-live/kiel-live" class="w-full">
+      </Button>
+      <Button href="https://github.com/kiel-live/kiel-live" class="w-full">
         <i-mdi-github class="mr-2" /><span>Develop on Github</span>
-      </CustomButton>
+      </Button>
 
       <h2 class="border-b-2 w-full dark:border-dark-100 mt-4 font-bold">Changelog</h2>
       <h3 class="border-b-1 dark:border-gray-600 mt-2">{{ new Date('2022-05-03').toLocaleDateString() }}</h3>
@@ -31,5 +31,5 @@
 </template>
 
 <script setup lang="ts">
-import CustomButton from '~/components/atomic/CustomButton.vue';
+import Button from '~/components/atomic/Button.vue';
 </script>
